pyproject.toml
[tool.poetry]
name = "asag-experiment"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = ">=3.11,<3.12"
pydantic = "^2.7.1"
loguru = "^0.7.2"
typer = "^0.12.3"
dspy-ai = "^2.4.9"
datasets = "2.14.6"
seaborn = "^0.13.2"
matplotlib = "^3.8.4"
pandas = "^2.2.2"
openai = "^1.30.1"
scikit-learn = "^1.4.2"
python-dotenv = "^1.0.1"


[tool.poetry.group.dev.dependencies]
black = "^24.4.2"
isort = "^5.13.2"
jupyter = "^1.0.0"
ipdb = "^0.13.13"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
